requirements.txt
streamlit>=1.28.0
anthropic>=0.7.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.17.0
requests>=2.31.0
yfinance>=0.2.0
schedule>=1.2.0
python-dotenv>=1.0.0